% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptonator.R
\name{Cryptonator.complete}
\alias{Cryptonator.complete}
\title{Cryptonator.complete}
\usage{
Cryptonator.complete(ticker = "btc-usd")
}
\arguments{
\item{ticker}{a single ticker (e.g. "btc-usd").  See
\code{\link{Cryptonator.currencies}} for a list of valid tickers}
}
\value{
a list with 2 components
\itemize{
  \item summary same information returned by \code{\link{Cryptonator.simple}}
  \item detail a data.frame with price and volume by exchange.
}
}
\description{
Returns actual volume-weighted price, total 24h volume, rate change as well
as prices and volumes across all connected exchanges.
}
\examples{
\dontrun{
Cryptonator.complete("btc-usd")
Cryptonator.complete("xmr-btc")
}
}
\author{
gsee
}

