% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptonator.R
\name{Cryptonator.currencies}
\alias{Cryptonator.currencies}
\title{Cryptonator.currencies}
\usage{
Cryptonator.currencies()
}
\value{
data.frame
\itemize{
 \item code currency code
 \item name description of currency
 \item statuses primary or secondary (or both)
}
}
\description{
Returns all the valid currency pairs
}
\examples{
\dontrun{
Cryptonator.currencies()
}
}

