% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptonator-package.R
\docType{package}
\name{cryptonatoR}
\alias{cryptonatoR}
\alias{cryptonatoR-package}
\title{cryptonatoR package: An \R interface to the Cryptonator(tm) REST API}
\description{
\emph{Cryptonator(tm) is a brand name that is owned by Global Technical
Enterprise Ltd.  This software is in no way affiliated, endorsed, or
approved by Cryptonator(tm), Global Technical Enterprise Ltd, or any of its
affiliates.}
}
\details{
The \pkg{cryptonatoR} package provides functions to access the free rate
quotes provided by Cryptonator(tm).
}
\examples{
\dontrun{
head(Cryptonator.currencies())
Cryptonator.simple("btc-usd")
Cryptonator.complete("btc-usd")
}
}
\author{
Garrett See \email{gsee000@gmail.com}
}
\references{
\url{https://www.cryptonator.com}

\url{https://www.cryptonator.com/api}
}
\keyword{IO}
\keyword{package}
\keyword{programming}

