% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptonator.R
\name{Cryptonator.simple}
\alias{Cryptonator.simple}
\title{Cryptonator.simple}
\usage{
Cryptonator.simple(ticker = "btc-usd")
}
\arguments{
\item{ticker}{a vector of tickers. Tickers are hyphen-separated
currency pairs (e.g. btc-usd, xmr-btc).  Note that if you provide more than
one ticker, a separate request will be made for each ticker.}
}
\value{
data.frame with
\itemize{
 \item ticker
 \item base base currency code
 \item target counter currency code
 \item price volume-weighted price
 \item volume total trade volume for the last 24 hours (Volume is displayed
 only for the cryptocurrencies that are actually traded on online exchanges.)
 \item change past hour price change
 \item timestamp POSIXct timestamp in local timezone
}
}
\description{
Returns actual volume-weighted price, total 24h volume and the price change
from Cryptonator.
}
\examples{
\dontrun{
Cryptonator.simple("btc-usd")
Cryptonator.simple(c("btc-usd", "xmr-btc", "xmr-usd"))
}
}
\author{
gsee
}

