% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cryptonator.R
\name{getQuote.cryptonator}
\alias{getQuote.cryptonator}
\title{getQuote.cryptonator}
\usage{
getQuote.cryptonator(Symbols, what = c("simple", "complete"))
}
\arguments{
\item{Symbols}{a ticker or vector of tickers (only supported for 
what="simple"). Tickers are hyphen-separated currency pairs (e.g. "btc-usd")}

\item{what}{either "simple" or "complete" depending on whether you want to 
get back a summary or details for all exchanges.}
}
\value{
see \code{\link{Cryptonator.simple}} or \code{\link{Cryptonator.complete}}
}
\description{
a "method" for \code{quantmod::getQuote}
}
\details{
This is simply a wrapper around \code{\link{Cryptonator.simple}} or 
 \code{\link{Cryptonator.complete}} to allow calling \code{quantmod::getQuote}
 with \code{src="cryptonator"}
}
\examples{
\dontrun{
library(quantmod)
getQuote("XMR-BTC", src="cryptonator")
}
}
\author{
gsee
}

